requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
Werkzeug==3.0.1
mysql-connector-python==8.3.0
pytz==2024.1 